+++ CartesAcces/frmParamtres.cs
@@ -1,16 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Drawing;
 using System.Drawing.Imaging;
 using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text.RegularExpressions;
 using System.Windows.Forms;
-using iText.Layout.Element;
-using iText.Kernel.Pdf.Xobject;
-using Rectangle = System.Drawing.Rectangle;
+using CartesAcces.Properties;
 
 namespace CartesAcces
 {
@@ -22,7 +17,7 @@ namespace CartesAcces
             Couleur.setCouleurFenetre(this);
             ControlSize.SetSizeTextControl(this);
             initDataGrid();
-            lblDateImport.Text = Properties.Settings.Default.DateImport;
+            lblDateImport.Text = Settings.Default.DateImport;
         }
 
         // -- Initialisation de la grille contenant la liste des élèves
@@ -35,15 +30,12 @@ namespace CartesAcces
 
         public void importEleves()
         {
-            string sourcePath = txtPathEleve.Text;
-            string destinationPath = Chemin.pathListeEleve;
+            var sourcePath = txtPathEleve.Text;
+            var destinationPath = Chemin.pathListeEleve;
             try
             {
-                if (File.Exists(destinationPath))
-                {
-                    File.Delete(destinationPath);
-                }
-                
+                if (File.Exists(destinationPath)) File.Delete(destinationPath);
+
                 Directory.CreateDirectory(Chemin.pathFolderListeEleve);
 
                 File.Copy(sourcePath, destinationPath);
@@ -53,34 +45,31 @@ namespace CartesAcces
                 initDataGrid();
                 MessageBox.Show("Import Réussi");
             }
-            catch(Exception e)
+            catch (Exception e)
             {
                 MessageBox.Show(e.ToString());
             }
-            
         }
 
         // -- Importation des photo des élèves --
         public void importPhoto()
         {
-            string sourcePath = txtPathPhoto.Text;
-            string destinationPath = Chemin.pathPhotoEleve;
+            var sourcePath = txtPathPhoto.Text;
+            var destinationPath = Chemin.pathPhotoEleve;
 
             try
             {
                 Directory.CreateDirectory(destinationPath);
-            
-                DirectoryInfo directory = new DirectoryInfo(sourcePath);
+
+                var directory = new DirectoryInfo(sourcePath);
 
                 foreach (var dir in directory.GetDirectories())
+                foreach (var file in dir.GetFiles())
                 {
-                    foreach (var file in dir.GetFiles())
-                    {
-                        System.Drawing.Image img = System.Drawing.Image.FromFile(file.FullName);
-                        string nom = file.Name;
+                    var img = Image.FromFile(file.FullName);
+                    var nom = file.Name;
 
-                        img.Save(destinationPath + nom, System.Drawing.Imaging.ImageFormat.Png);
-                    }
+                    img.Save(destinationPath + nom, ImageFormat.Png);
                 }
 
                 MessageBox.Show("Import réussie !");
@@ -89,24 +78,20 @@ namespace CartesAcces
             {
                 MessageBox.Show(e.ToString());
             }
-
-
         }
 
         // -- La liste des élève se met a jour a chaques saisies de l'utilisateur --
         public void rechercheDataGrid()
         {
-            List<Eleve> listeRecherche = new List<Eleve>();
+            var listeRecherche = new List<Eleve>();
 
             initDataGrid();
 
             foreach (DataGridViewRow row in DataGridParametres.Rows)
-            {
-                for (int i = 0; i < row.Cells.Count; i++)
-                {
+                for (var i = 0; i < row.Cells.Count; i++)
                     if (row.Cells[i].Value.ToString().Contains(txtRechercheDataGrid.Text))
                     {
-                        Eleve eleve = new Eleve();
+                        var eleve = new Eleve();
 
                         eleve.NomEleve = row.Cells[0].Value.ToString();
                         eleve.PrenomEleve = row.Cells[1].Value.ToString();
@@ -120,15 +105,14 @@ namespace CartesAcces
 
                         listeRecherche.Add(eleve);
                     }
-                }
-            }
 
             if (listeRecherche.Count() == 0)
             {
                 MessageBox.Show("Unable to find " + txtRechercheDataGrid.Text, "Not Found");
                 return;
             }
-            else if (txtRechercheDataGrid.Text == "")
+
+            if (txtRechercheDataGrid.Text == "")
             {
                 initDataGrid();
                 listeRecherche.Clear();
@@ -144,18 +128,13 @@ namespace CartesAcces
         // -- Pour la photo uniques --
         public string getImportPath()
         {
-            using (OpenFileDialog ofd = new OpenFileDialog())
+            using (var ofd = new OpenFileDialog())
             {
                 ofd.Filter = "Jpeg Files Only | *.jpg";
                 ofd.Title = "Selectionner une photo";
                 if (ofd.ShowDialog() == DialogResult.OK)
-                {
                     return ofd.FileName;
-                }
-                else
-                {
-                    return "null";
-                }
+                return "null";
             }
         }
 
@@ -164,24 +143,21 @@ namespace CartesAcces
         private void btnImporterEleves_Click(object sender, EventArgs e)
         {
             txtPathEleve.Text = Chemin.setPathImportFileEXCEL();
-            if (txtPathEleve.Text.Length > 0)
-            {
-                btnValiderEleve.Enabled = true;
-            }
+            if (txtPathEleve.Text.Length > 0) btnValiderEleve.Enabled = true;
         }
 
         private void btnValiderEleve_Click(object sender, EventArgs e)
         {
             labelV.Show();
             importEleves();
-            lblDateImport.Text = Properties.Settings.Default.DateImport;
+            lblDateImport.Text = Settings.Default.DateImport;
             labelV.Hide();
         }
 
         private void btnImportEDT_Click(object sender, EventArgs e)
         {
             txtPathEDT.Text = Chemin.setPathImportFilePDF();
-            frmSelectSection frmSelectSection = new frmSelectSection();
+            var frmSelectSection = new frmSelectSection();
             //frmSelectSection.Top = new frmParametres().Top;
             frmSelectSection.StartPosition = FormStartPosition.CenterScreen;
             frmSelectSection.Show();
@@ -206,17 +182,14 @@ namespace CartesAcces
 
         private void btnAjoutEleve_Click(object sender, EventArgs e)
         {
-
         }
 
         private void btnModifClasse_Click(object sender, EventArgs e)
         {
-
         }
 
         private void btnSupprEleve_Click(object sender, EventArgs e)
         {
-
         }
 
         private void txtRechercheDataGrid_TextChanged(object sender, EventArgs e)
@@ -226,22 +199,20 @@ namespace CartesAcces
 
         private void DataGridParametres_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
-
         }
 
         private void frmParametres_Load(object sender, EventArgs e)
         {
-            Timer time = new Timer(this);
+            var time = new Timer(this);
         }
 
         private void label2_Click(object sender, EventArgs e)
         {
-
         }
 
         private void btnImportPhotoUnique_Click(object sender, EventArgs e)
         {
-            string path = getImportPath();
+            var path = getImportPath();
 
             if (path == "null")
             {
@@ -249,7 +220,7 @@ namespace CartesAcces
                 return;
             }
 
-            frmImportPhotoUnique frm = new frmImportPhotoUnique();
+            var frm = new frmImportPhotoUnique();
 
             //Photo.setLaPhoto(path, );
